refactor: do not draw layers in batch segmentation feature: add remove_border_touching_components() also for batch segmentation
