+++ src/napari_roxas_ai/_segmentation/_single_sample_segmentation.py
@@ -25,6 +25,7 @@ from napari_roxas_ai._settings import SettingsManager
 from napari_roxas_ai._utils import make_binary_labels_colormap
 
 from ._cells_model import CellsSegmentationModel
+from .._utils._segmentation_postprocess import remove_border_touching_components
 
 if TYPE_CHECKING:
     import napari
@@ -181,50 +182,6 @@ def apply_segmentation_results_to_viewer(
             rings_layer.features = new_rings_table
             rings_layer.data = new_rings_raster
             rings_layer.colormap = new_colormap
-
-
-def remove_border_touching_components(binary: np.ndarray) -> np.ndarray:
-    """
-    Remove connected components that touch the image border.
-    Expects a 2D binary array (0/1 or 0/255).
-    Returns a uint8 array with values 0/1.
-    """
-    if binary.ndim != 2:
-        raise ValueError(f"Expected 2D array, got shape={binary.shape}")
-
-    mask = (binary > 0)
-
-    if not mask.any():
-        return mask.astype("uint8")
-
-    # 8-connectivity for 2D
-    structure = np.ones((3, 3), dtype=bool)
-    labeled, num = ndi.label(mask, structure=structure)
-
-    if num == 0:
-        return mask.astype("uint8")
-
-    # Collect labels that appear on the border
-    border_labels = np.unique(
-        np.concatenate(
-            [
-                labeled[0, :],
-                labeled[-1, :],
-                labeled[:, 0],
-                labeled[:, -1],
-            ]
-        )
-    )
-    border_labels = border_labels[border_labels != 0]  # ignore background
-
-    if border_labels.size == 0:
-        return mask.astype("uint8")
-
-    # Remove all pixels belonging to those labels
-    remove = np.isin(labeled, border_labels)
-    mask[remove] = False
-
-    return mask.astype("uint8")
 
 
 class Worker(QObject):
